Bug Fixes and General Updates
Fixed a Few Bugs and updated some data. The last thing that i need to do
is provide a better implementation for the pulling and pushing of data.
Its currently hard wired.

That change should be implemented by tomorrow.

Class Diagram was created and placed into the class diagram folder.

Tab order fixed.

What needs to be done:

Documentation & Comments
Proper Implementation of Push and Pull (Just need to fix the push)